feat: Continue integration and download of more artifacts

+++ ui/download.go
@@ -345,11 +345,8 @@ func (s *DownloadScreen) buildDownloads(config internal.Config, host romm.Host, 
 
 	for _, g := range games {
 		gamelistRomEntry := gamelist.RomGameEntry{
-			Game:         &g,
-			ArtLocation:  "",
-			GamePath:     "",
-			RomDirectory: "",
-			Platform:     &platform,
+			Game:     &g,
+			Platform: &platform,
 		}
 		gamePlatform := platform
 		if platform.ID == 0 && g.PlatformID != 0 {
@@ -402,7 +399,7 @@ func (s *DownloadScreen) buildDownloads(config internal.Config, host romm.Host, 
 			artLocation := filepath.Join(artDir, artFileName)
 
 			coverURL := g.GetArtworkURL(config.ArtKind, host)
-			gamelistRomEntry.ArtLocation = artLocation
+			gamelistRomEntry.ArtLocation.ImagePath = artLocation
 
 			artDownloads = append(artDownloads, artDownload{
 				URL:      coverURL,
@@ -426,9 +423,23 @@ func (s *DownloadScreen) buildDownloads(config internal.Config, host romm.Host, 
 			}
 
 			artSplashDir := config.GetArtSplashDirectory(gamePlatform)
-			if config.DownloadSplashArt != artutil.ArtKindNone && artSplashDir != "" {
-				splashArtLocation := filepath.Join(artSplashDir, artFileName)
-				if splashURL := g.GetSplashArtURL(config.DownloadSplashArt, host); splashURL != "" {
+			if (config.DownloadSplashArt != artutil.ArtKindNone || config.AdditionalDownloads.Thumbnail != artutil.ArtKindNone) && artSplashDir != "" {
+				artSplashFileName := g.FsNameNoExt
+				isESBased := cfw.GetCFW().IsBasedOnEmulationStation()
+				if isESBased {
+					artSplashFileName += "-thumb.png"
+				} else {
+					artSplashFileName += ".png"
+				}
+				splashArtLocation := filepath.Join(artSplashDir, artSplashFileName)
+				kind := config.DownloadSplashArt
+				if config.AdditionalDownloads.Thumbnail != artutil.ArtKindNone {
+					kind = config.AdditionalDownloads.Thumbnail
+				}
+				if splashURL := g.GetSplashArtURL(kind, host); splashURL != "" {
+					if isESBased {
+						gamelistRomEntry.ArtLocation.ThumbnailPath = splashArtLocation
+					}
 					artDownloads = append(artDownloads, artDownload{
 						URL:      splashURL,
 						Location: splashArtLocation,
@@ -437,6 +448,119 @@ func (s *DownloadScreen) buildDownloads(config internal.Config, host romm.Host, 
 					})
 				}
 			}
+
+			artMarqueeDir := config.GetArtMarqueeDirectory(gamePlatform)
+			if config.AdditionalDownloads.Marquee != artutil.ArtKindNone && artMarqueeDir != "" {
+				marqueeArtFileName := g.FsNameNoExt
+				// is cfw is ES based, use -marquee suffix to avoid conflicts with cover art
+				if cfw.GetCFW().IsBasedOnEmulationStation() {
+					marqueeArtFileName += "-marquee.png"
+				} else {
+					marqueeArtFileName += ".png"
+				}
+				marqueeArtLocation := filepath.Join(artMarqueeDir, marqueeArtFileName)
+				marqueeURL := ""
+				switch config.AdditionalDownloads.Marquee {
+				case artutil.ArtKindMarquee:
+					marqueeURL = g.GetMarqueeURL(host)
+				case artutil.ArtKindLogo:
+					marqueeURL = g.GetLogoURL(host)
+				}
+				if marqueeURL != "" {
+					gamelistRomEntry.ArtLocation.MarqueePath = marqueeArtLocation
+					artDownloads = append(artDownloads, artDownload{
+						URL:      marqueeURL,
+						Location: marqueeArtLocation,
+						GameName: g.Name,
+						IsImage:  true,
+					})
+				}
+			}
+
+			artVideoDir := config.GetArtVideoDirectory(gamePlatform)
+			if config.AdditionalDownloads.Video && artVideoDir != "" {
+				videoLocation := filepath.Join(artVideoDir, g.FsNameNoExt+".mp4")
+				if videoURL := g.GetVideoURL(host); videoURL != "" {
+					gamelistRomEntry.ArtLocation.VideoPath = videoLocation
+					artDownloads = append(artDownloads, artDownload{
+						URL:      videoURL,
+						Location: videoLocation,
+						GameName: g.Name,
+						IsImage:  false,
+					})
+				}
+			}
+
+			artBezelDir := config.GetArtBezelDirectory(gamePlatform)
+			if config.AdditionalDownloads.Bezel && artBezelDir != "" {
+				bezelArtLocation := filepath.Join(artBezelDir, artFileName)
+				if bezelURL := g.GetBezelURL(host); bezelURL != "" {
+					gamelistRomEntry.ArtLocation.BezelPath = bezelArtLocation
+					artDownloads = append(artDownloads, artDownload{
+						URL:      bezelURL,
+						Location: bezelArtLocation,
+						GameName: g.Name,
+						IsImage:  true,
+					})
+				}
+			}
+
+			manualDir := config.GetManualDirectory(gamePlatform)
+			if config.AdditionalDownloads.Manual && manualDir != "" {
+				manualLocation := filepath.Join(manualDir, g.FsNameNoExt+".pdf")
+				if manualURL := g.GetManualURL(host); manualURL != "" {
+					gamelistRomEntry.ArtLocation.ManualPath = manualLocation
+					artDownloads = append(artDownloads, artDownload{
+						URL:      manualURL,
+						Location: manualLocation,
+						GameName: g.Name,
+						IsImage:  false,
+					})
+				}
+			}
+
+			boxbackDir := config.GetBoxbackDirectory(gamePlatform)
+			if config.AdditionalDownloads.BoxBack && boxbackDir != "" {
+				boxbackArtFileName := g.FsNameNoExt
+				// is cfw is ES based, use -boxback suffix to avoid conflicts with cover art
+				if cfw.GetCFW().IsBasedOnEmulationStation() {
+					boxbackArtFileName += "-boxback.png"
+				} else {
+					boxbackArtFileName += ".png"
+				}
+				boxbackArtLocation := filepath.Join(boxbackDir, boxbackArtFileName)
+				if boxbackURL := g.GetBoxbackURL(host); boxbackURL != "" {
+					gamelistRomEntry.ArtLocation.BoxBackPath = boxbackArtLocation
+					artDownloads = append(artDownloads, artDownload{
+						URL:      boxbackURL,
+						Location: boxbackArtLocation,
+						GameName: g.Name,
+						IsImage:  true,
+					})
+				}
+			}
+
+			fanartDir := config.GetFanartDirectory(gamePlatform)
+			if config.AdditionalDownloads.Fanart && fanartDir != "" {
+				fanartFileName := g.FsNameNoExt
+				// is cfw is ES based, use -fanart suffix to avoid conflicts with cover art
+				if cfw.GetCFW().IsBasedOnEmulationStation() {
+					fanartFileName += "-fanart.png"
+				} else {
+					fanartFileName += ".png"
+				}
+				fanartLocation := filepath.Join(fanartDir, fanartFileName)
+				if fanartURL := g.GetFanartURL(host); fanartURL != "" {
+					gamelistRomEntry.ArtLocation.FanartPath = fanartLocation
+					artDownloads = append(artDownloads, artDownload{
+						URL:      fanartURL,
+						Location: fanartLocation,
+						GameName: g.Name,
+						IsImage:  true,
+					})
+				}
+			}
+
 		}
 		gamesSummaries = append(gamesSummaries, gamelistRomEntry)
 	}
@@ -549,15 +673,17 @@ func (s *DownloadScreen) downloadArt(artDownloads []artDownload, downloadedGames
 			continue
 		}
 
-		if err := imageutil.ProcessArtImage(art.Location); err != nil {
-			logger.Warn("Failed to process art image", "game", art.GameName, "location", art.Location, "error", err)
-			os.Remove(art.Location)
-			failCount++
-			processedCount++
-			if totalArt > 0 {
-				progress.Store(float64(processedCount) / float64(totalArt))
+		if art.IsImage {
+			if err := imageutil.ProcessArtImage(art.Location); err != nil {
+				logger.Warn("Failed to process art image", "game", art.GameName, "location", art.Location, "error", err, "url", art.URL)
+				os.Remove(art.Location)
+				failCount++
+				processedCount++
+				if totalArt > 0 {
+					progress.Store(float64(processedCount) / float64(totalArt))
+				}
+				continue
 			}
-			continue
 		}
 
 		successCount++
@@ -567,5 +693,4 @@ func (s *DownloadScreen) downloadArt(artDownloads []artDownload, downloadedGames
 			progress.Store(float64(processedCount) / float64(totalArt))
 		}
 	}
-
 }
